refactor: rename `sum_up` module to `sum` and update imports accordingly
